Fixed issues with displaying SVG background maps. Added Peter Mumford 18GA redesigned map. (cherry picked from commit 7cf64e84042d36e1f0c1d902e767efc987f797be)

+++ rails/ui/swing/hexmap/GUIHex.java
@@ -93,7 +93,7 @@ public class GUIHex implements ViewObject {
     protected static Logger log =
             Logger.getLogger(GUIHex.class.getPackage().getName());
 
-    public GUIHex(HexMap hexMap, double cx, double cy, int scale,
+    public GUIHex(HexMap hexMap, double cx, double cy, double scale,
             int xCoord, int yCoord) {
         this.hexMap = hexMap;
         this.x = xCoord;
@@ -103,7 +103,7 @@ public class GUIHex implements ViewObject {
 
     }
 
-    public void scaleHex (double cx, double cy, int scale, double zoomFactor) {
+    public void scaleHex (double cx, double cy, double scale, double zoomFactor) {
 
         this.zoomFactor = zoomFactor;
         tokenDiameter = (int)Math.round(NORMAL_TOKEN_SIZE * zoomFactor);
